chore(team): split step continuity helpers
